Pulling shared fields/accessors up to OperationParameters superclass
